+++ env.yml
@@ -1,8 +1,6 @@
 channels:
   - defaults
   - conda-forge
-  - pytorch  # TODO (necessary?)
-  - nvidia  # TODO (necessary?)
 dependencies:
   - python=3.12
   - transformers
@@ -16,7 +14,6 @@ dependencies:
   - dvc-s3
   - cattrs
   - attrs
-  - plotly
   - statsmodels
   - biotite
 
@@ -39,3 +36,4 @@ dependencies:
       - nbsphinx==0.9.5
       - nbsphinx-link==1.3.1
       - sphinx-comments==0.0.3
+      - plotly
