+++ src/analysis/__init__.py
@@ -1,0 +1,8 @@
+"""
+This package contains all the code we used in the analysis that we didn't include inline
+within the notebook. The thinking here is that we don't want to cloud or bloat the
+notebook with code that doesn't significantly contribute to reader understanding. Of
+course, this code is still critical for reproducibilty, and so the documentation for
+this code is included herein, and the source code can be read directly by clicking the
+"[source]" button associated with each object.
+"""
